fix 'add book to cart'

+++ public/js/view/buyer/ebook_info.js
@@ -304,15 +304,26 @@ function addBookToCart() {
   $('#btnAddToCart').attr("onclick", "");
   $('#btnAddToCart').html("....");
   var bookId = $('meta[name=book-id]').attr("content");
-  $.ajax({
-    url : "/post/add_book_to_cart/"+bookId,
-    method : "POST"
-  }).done(function() {
-    $('#btnAddToCart').html("Hapus dari Keranjang");
-    $('#btnAddToCart').attr("onclick", "deleteBookFromCart()");
-    $('#btnAddToCart').attr("id", "btnDeleteFromCart");
-    storeFlashMessage("Berhasil menambah ebook ke keranjang belanja", "success", 2);
-    setUpAlert();
+  $.ajax({ // mengecek apakah buku sudah dihapus
+    url : '/get/is_book_not_deleted',
+    method : "GET",
+    data : { "id" : bookId }
+  }).done(function(isNotDeleted) {
+    if (isNotDeleted) { // jika blm dihapus
+      $.ajax({ // memasukkan buku ke keranjang
+        url : "/post/add_book_to_cart/"+bookId,
+        method : "POST"
+      }).done(function() {
+        $('#btnAddToCart').html("Hapus dari Keranjang");
+        $('#btnAddToCart').attr("onclick", "deleteBookFromCart()");
+        $('#btnAddToCart').attr("id", "btnDeleteFromCart");
+        storeFlashMessage("Berhasil menambah ebook ke keranjang belanja", "success", 2);
+        setUpAlert();
+      });
+    }
+    else {
+      location.reload();
+    }
   });
 }
